requirements.txt
SpeechRecognition==3.14.5
PyAudio==0.2.14
requests==2.32.5
playsound==1.2.2
edge-tts==7.2.7
python-dotenv==1.2.1
groq==1.0.0
